Remove unused image assets and add placeholder images; Moves design system colors and fonts to App folder
