--- Freela-onTap/App/Assets.xcassets/companyPhotos/Contents.json
@@ -1,9 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "provides-namespace" : true
-  }
-}
