+++ Freela-onTap/App/DesignSystem/UIColor+DesignSystem.swift
@@ -1,0 +1,54 @@
+//
+//  UIColor+DesignSystem.swift
+//  App-Challenge
+//
+//
+
+import UIKit
+
+extension UIColor {
+    /// Xcode's autocomplete allows for easy discovery of design system UIColors.
+    /// At any call site that requires a UIColor, type `UIColor.DesignSystem.<esc>`
+    struct DesignSystem {
+        static let lavanda0 = UIColor(red: 0.9647058844566345, green: 0.9607843160629272, blue: 1, alpha: 1)
+        static let lavanda100 = UIColor(red: 0.800000011920929, green: 0.772549033164978, blue: 0.9921568632125854, alpha: 1)
+        static let lavanda200 = UIColor(red: 0.6392157077789307, green: 0.5960784554481506, blue: 0.9764705896377563, alpha: 1)
+        static let lavanda300 = UIColor(red: 0.49803921580314636, green: 0.43921568989753723, blue: 0.9490196108818054, alpha: 1)
+        static let lavanda400 = UIColor(red: 0.3843137323856354, green: 0.3137255012989044, blue: 0.9137254953384399, alpha: 1)
+        static let lavanda500 = UIColor(red: 0.2980392277240753, green: 0.21960784494876862, blue: 0.8666666746139526, alpha: 1)
+        static let lavanda600 = UIColor(red: 0.22745098173618317, green: 0.15294118225574493, blue: 0.8039215803146362, alpha: 1)
+        static let lavanda700 = UIColor(red: 0.1764705926179886, green: 0.10588235408067703, blue: 0.7215686440467834, alpha: 1)
+        static let lavanda800 = UIColor(red: 0.13725490868091583, green: 0.07450980693101883, blue: 0.6274510025978088, alpha: 1)
+        static let lavanda900 = UIColor(red: 0.10980392247438431, green: 0.054901961237192154, blue: 0.529411792755127, alpha: 1)
+        static let terracota0 = UIColor(red: 1, green: 1, blue: 1, alpha: 1)
+        static let terracota100 = UIColor(red: 0.9921568632125854, green: 0.8313725590705872, blue: 0.772549033164978, alpha: 0.30000001192092896)
+        static let terracota200 = UIColor(red: 0.9764705896377563, green: 0.6941176652908325, blue: 0.5960784554481506, alpha: 1)
+        static let terracota300 = UIColor(red: 0.9490196108818054, green: 0.5764706134796143, blue: 0.43921568989753723, alpha: 1)
+        static let terracota400 = UIColor(red: 0.9137254953384399, green: 0.4745098054409027, blue: 0.3137255012989044, alpha: 1)
+        static let terracota500 = UIColor(red: 0.8666666746139526, green: 0.3921568691730499, blue: 0.21960784494876862, alpha: 1)
+        static let terracota600 = UIColor(red: 0.8039215803146362, green: 0.32549020648002625, blue: 0.15294118225574493, alpha: 1)
+        static let terracota700 = UIColor(red: 0.7215686440467834, green: 0.2705882489681244, blue: 0.10588235408067703, alpha: 1)
+        static let terracota800 = UIColor(red: 0.6274510025978088, green: 0.21960784494876862, blue: 0.07450980693101883, alpha: 1)
+        static let terracota900 = UIColor(red: 0.529411792755127, green: 0.18039216101169586, blue: 0.054901961237192154, alpha: 1)
+        static let petroleo0 = UIColor(red: 0.9607843160629272, green: 0.9882352948188782, blue: 1, alpha: 1)
+        static let petroleo100 = UIColor(red: 0.772549033164978, green: 0.9215686321258545, blue: 0.9921568632125854, alpha: 1)
+        static let petroleo200 = UIColor(red: 0.5960784554481506, green: 0.8549019694328308, blue: 0.9764705896377563, alpha: 1)
+        static let petroleo300 = UIColor(red: 0.43921568989753723, green: 0.7882353067398071, blue: 0.9490196108818054, alpha: 1)
+        static let petroleo400 = UIColor(red: 0.3137255012989044, green: 0.7254902124404907, blue: 0.9137254953384399, alpha: 1)
+        static let petroleo500 = UIColor(red: 0.21960784494876862, green: 0.6627451181411743, blue: 0.8666666746139526, alpha: 1)
+        static let petroleo600 = UIColor(red: 0.15294118225574493, green: 0.5960784554481506, blue: 0.8039215803146362, alpha: 1)
+        static let petroleo700 = UIColor(red: 0.10588235408067703, green: 0.5254902243614197, blue: 0.7215686440467834, alpha: 1)
+        static let petroleo800 = UIColor(red: 0.07450980693101883, green: 0.45098039507865906, blue: 0.6274510025978088, alpha: 1)
+        static let petroleo900 = UIColor(red: 0.054901961237192154, green: 0.3803921639919281, blue: 0.529411792755127, alpha: 1)
+        static let azulao0 = UIColor(red: 0.9607843160629272, green: 0.9803921580314636, blue: 1, alpha: 1)
+        static let azulao100 = UIColor(red: 0.772549033164978, green: 0.8784313797950745, blue: 0.9921568632125854, alpha: 1)
+        static let azulao200 = UIColor(red: 0.5960784554481506, green: 0.7803921699523926, blue: 0.9764705896377563, alpha: 1)
+        static let azulao300 = UIColor(red: 0.43921568989753723, green: 0.686274528503418, blue: 0.9490196108818054, alpha: 1)
+        static let azulao400 = UIColor(red: 0.3137255012989044, green: 0.6039215922355652, blue: 0.9137254953384399, alpha: 1)
+        static let azulao500 = UIColor(red: 0.21960784494876862, green: 0.5333333611488342, blue: 0.8666666746139526, alpha: 1)
+        static let azulao600 = UIColor(red: 0.15294118225574493, green: 0.46666666865348816, blue: 0.8039215803146362, alpha: 1)
+        static let azulao700 = UIColor(red: 0.10588235408067703, green: 0.40392157435417175, blue: 0.7215686440467834, alpha: 1)
+        static let azulao800 = UIColor(red: 0.07450980693101883, green: 0.34117648005485535, blue: 0.6274510025978088, alpha: 1)
+        static let azulao900 = UIColor(red: 0.054901961237192154, green: 0.2823529541492462, blue: 0.529411792755127, alpha: 1)
+    }
+}
